# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LARAVEL Rest API DOCUMENTATION",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "manirujjamanakash@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/movies/{id}": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Get Detail",
                "description": "Get Detail",
                "operationId": "03fb9b29d3c7a8089f4b5c1d693c3172",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "movie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Movie"
                ],
                "summary": "Movie Update",
                "description": "Update Movie",
                "operationId": "cbe81c76a99cad2d227684e276a91c6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "movie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Movie objects",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "title",
                                        "type": "string",
                                        "example": "love"
                                    },
                                    "summary": {
                                        "description": "summary",
                                        "type": "string",
                                        "example": "about fight"
                                    },
                                    "cover_image": {
                                        "description": "cover_image",
                                        "type": "file",
                                        "example": "image"
                                    },
                                    "imdb_ratings": {
                                        "description": "imdb_ratings",
                                        "type": "string",
                                        "example": "9"
                                    },
                                    "publish_status": {
                                        "description": "publish_status",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "pdf_download_link": {
                                        "description": "pdf_download_link",
                                        "type": "string",
                                        "example": "www.google.com"
                                    },
                                    "genre_ids": {
                                        "description": "genre_ids",
                                        "type": "string",
                                        "example": "[1,2]"
                                    },
                                    "tag_ids": {
                                        "description": "tag_ids",
                                        "type": "string",
                                        "example": "[1,2]"
                                    },
                                    "author_ids": {
                                        "description": "author_ids",
                                        "type": "string",
                                        "example": "[1]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Movie"
                ],
                "summary": "Delete",
                "description": "Delete",
                "operationId": "0f4a98c2e92b423a382c24e562063d0e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Movie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully Deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Level not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/movies/related-movies/{id}": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Get relatedMovie",
                "description": "Get relatedMovie",
                "operationId": "fefbc4c7c5ce8ba212576503673852b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "movie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/movies": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Get MovieList",
                "description": "Get MovieList",
                "operationId": "1370ed35be39cf7de2c9fb83b00118d2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Movie"
                ],
                "summary": "Movie Create",
                "description": "Create Movie",
                "operationId": "add6773ac03f9fab7aac9eb459ee960b",
                "requestBody": {
                    "description": "Movie objects",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "title",
                                        "type": "string",
                                        "example": "love"
                                    },
                                    "summary": {
                                        "description": "summary",
                                        "type": "string",
                                        "example": "about fight"
                                    },
                                    "cover_image": {
                                        "description": "cover_image",
                                        "type": "file",
                                        "example": "image"
                                    },
                                    "imdb_ratings": {
                                        "description": "imdb_ratings",
                                        "type": "string",
                                        "example": "9"
                                    },
                                    "publish_status": {
                                        "description": "publish_status",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "pdf_download_link": {
                                        "description": "pdf_download_link",
                                        "type": "string",
                                        "example": "www.google.com"
                                    },
                                    "genre_ids": {
                                        "description": "genre_ids",
                                        "type": "string",
                                        "example": "[1,2]"
                                    },
                                    "tag_ids": {
                                        "description": "tag_ids",
                                        "type": "string",
                                        "example": "[1,2]"
                                    },
                                    "author_ids": {
                                        "description": "author_ids",
                                        "type": "string",
                                        "example": "[1]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "Enter your bearer token in the format `Bearer {token}`"
            }
        }
    }
}